refactoring
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2334 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/assembly64/ZeroContainingRatingConverter.java
@@ -19,7 +19,7 @@ public class ZeroContainingRatingConverter extends StringConverter<Integer> {
 
 	@Override
 	public String toString(Integer rating) {
-		if (rating == 0) {
+		if (rating.intValue() == 0) {
 			return bundle.getString("ALL_CONTENT");
 		}
 		return String.valueOf(rating);
